Refactor in BookController

+++ www/open-library/app/Http/Controllers/BookController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Author;
 use App\Models\Book;
+use App\Models\Author;
+use App\Models\Category;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
-use App\Validation\IsbnValidator;
 use Nicebooks\Isbn\IsbnTools;
+use App\Validation\IsbnValidator;
+use Illuminate\Support\Facades\Storage;
+use PhpParser\Node\Stmt\TryCatch;
 
 class BookController extends Controller
 {
@@ -30,6 +32,7 @@ class BookController extends Controller
     public function create()
     {
         $data['authors'] = Author::all();
+        $data['categories'] = Category::all();
         return view('book.create', $data);
     }
 
@@ -41,38 +44,7 @@ class BookController extends Controller
      */
     public function store(Request $request)
     {
-        $isbnFormatter = new IsbnTools;
-        //validate data
-        $fields = [
-            'isbn' => ['required', 'string', new IsbnValidator],
-            'title' => 'required|string|max:255',
-            'publication_year' => 'int|max:99999|nullable',
-            'cover_image' => 'max:100000|mimes:jpeg,png,jpg|nullable'
-        ];
-
-        $errorMessage = [
-            'required' => 'The :attribute is required',
-        ];
-
-        $this->validate($request, $fields, $errorMessage);
-
-        $bookData = $request->except('_token');
-
-        if ($request->has('authors')) {
-            //make author relationships
-            $authors = $request->input('authors');
-            Book::author()->sync($authors);
-        }
-
-        if ($request->hasFile('cover_image')) {
-            $bookData['cover_image'] = $request->file('cover_image')->store('uploads', 'public');
-        }
-
-        $bookData['isbn'] = $isbnFormatter->format($bookData['isbn']);
-
-        Book::insert($bookData);
-
-        return redirect('book')->with('message', 'Book created successfully');
+        return $this->insertOrUpdate($request);
     }
 
     /**
@@ -95,7 +67,41 @@ class BookController extends Controller
     public function edit($bookId)
     {
         $book = Book::findOrFail($bookId);
-        return view('book.edit', compact('book'));
+
+        $data = self::populateRelatedFields();
+        $data['book'] = $book;
+        $data['selected_authors'] = self::getSelectedAuthorsIds($book);
+        $data['selected_categories'] = self::getSelectedCategoriesIds($book);
+        //return view('book.edit', compact('book'));
+        return view('book.edit', $data);
+    }
+
+    public static function populateRelatedFields()
+    {
+        $data['authors'] = Author::all();
+        $data['categories'] = Category::all();
+        return $data;
+    }
+    public static function getSelectedAuthorsIds(Book $book)
+    {
+        $selected_authors = [];
+        if (!empty($book->authors)) {
+            foreach ($book->authors as $author) {
+                $selected_authors[] = $author->id;
+            }
+        }
+        return $selected_authors;
+    }
+
+    public static function getSelectedCategoriesIds(Book $book)
+    {
+        $selected_categories = [];
+        if (!empty($book->categories)) {
+            foreach ($book->categories as $category) {
+                $selected_categories[] = $category->id;
+            }
+        }
+        return $selected_categories;
     }
 
     /**
@@ -107,8 +113,35 @@ class BookController extends Controller
      */
     public function update(Request $request, $bookId)
     {
-        //Get the current data to make some checks with existing values
-        $currentBook = Book::findOrFail($bookId);
+        return $this->insertOrUpdate($request, $bookId);
+    }
+
+    /**
+     * Remove the specified book from storage.
+     *
+     * @param  Int $bookId
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy($bookId)
+    {
+        $book = Book::findOrFail($bookId);
+        if (Storage::delete('public/' . $book->cover_image)) {
+            //Check if the image could be deleted amd then delete the Book
+            Book::destroy($bookId);
+        } else {
+            return redirect('book/' . $book->id . '/edit')->with('error', 'Error while deleting the book cover, please try again later. If the problem persists contact an administrator');
+        }
+        return redirect('book')->with('message', 'Book deleted successfully');
+    }
+
+    private function insertOrUpdate($request, $updateId = false)
+    {
+        $currentBook = NULL;
+        if (!empty($updateId)) {
+            //Get the current data to make some checks with existing values
+            $currentBook = Book::findOrFail($updateId);
+        }
+
 
         $isbnFormatter = new IsbnTools;
         //validate data
@@ -125,44 +158,53 @@ class BookController extends Controller
 
         $this->validate($request, $fields, $errorMessage);
 
-        //Exclude useless fields
-        $bookData = $request->except(['_token', '_method']);
+        $bookData = $request->except(['_token', '_method', 'authors', 'categories']);
 
-        //If an image is received then update the cover_image
         if ($request->hasFile('cover_image')) {
-            if (!empty($currentBook->cover_image)) {
+            if (!empty($updateId)) {
+                //If the request is update we have to make some checks with images
+                //If an image is received then update the cover_image
                 //Check if an image already exists and delete it
-                Storage::delete('public/' . $currentBook->cover_image);
+                if (!empty($currentBook->cover_image)) {
+                    Storage::delete('public/' . $currentBook->cover_image);
+                }
             }
             $bookData['cover_image'] = $request->file('cover_image')->store('uploads', 'public');
         }
+
         //Check if th isbn has changed
-        if ($currentBook->isbn != $request->isbn) {
+        if (empty($updateId) || $currentBook->isbn != $request->isbn) {
             //The isbn has changed so we need to filter it again
             $bookData['isbn'] = $isbnFormatter->format($bookData['isbn']);
         }
-        //Update with new data
-        Book::where('id', '=', $bookId)->update($bookData);
-        $book = Book::findOrFail($bookId);
 
-        return redirect('book')->with('message', 'Book updated successfully');
-    }
 
-    /**
-     * Remove the specified book from storage.
-     *
-     * @param  Int $bookId
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($bookId)
-    {
-        $book = Book::findOrFail($bookId);
-        if (Storage::delete('public/' . $book->cover_image)) {
-            //Check if the image could be deleted amd then delete the Book
-            Book::destroy($bookId);
-        } else {
-            return redirect('book/' . $book->id)->with('error', 'Error while deleting the book cover, please try again later. If the problem persists contact an administrator');
+        //If we are updating, then get the existing book to be edited
+        $book = empty($updateId) ? new Book() : Book::findOrFail($updateId);
+
+        //Iterate through the fields and set thir values
+        foreach ($bookData as $field => $value) {
+            $book->$field = $value;
         }
-        return redirect('book')->with('message', 'Book deleted successfully');
+        //Save the book an then add relationships
+        $book->save();
+
+        if ($request->has('categories')) {
+            //make categories relationships
+            $categories = $request->input('categories');
+            $book->categories()->sync($categories);
+        }
+
+        if ($request->has('authors')) {
+            //make author relationships
+            $authors = $request->input('authors');
+            $book->authors()->sync($authors);
+        }
+
+
+
+        $action = empty($updateId) ? 'created' : 'updated';
+
+        return redirect('book/')->with('message', 'Book ' . $action . ' successfully');
     }
 }
